feat: grouping Toolbar buttons under the 'Edit' dropdown button (#3608)
* let ToolBar support 'dropdown' type

* move 'design' page only toolbar items into DesignPage

* migrate 'Add New Trigger' coachmark feature

* grouping 'Edit' buttons together as a dropdown

* loosen 'type' TS type in toolbar
