+++ Composer/packages/client/src/components/ToolBar/index.ts
@@ -1,0 +1,5 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+export { ToolBar } from './ToolBar';
+export { IToolBarItem } from './ToolBar.types';
