+++ Composer/packages/client/src/pages/design/DesignPage.tsx
@@ -3,7 +3,7 @@
 
 /** @jsx jsx */
 import { jsx } from '@emotion/core';
-import React, { Suspense, useEffect, useMemo, useState } from 'react';
+import React, { Suspense, useEffect, useMemo, useState, useCallback } from 'react';
 import { Breadcrumb, IBreadcrumbItem } from 'office-ui-fabric-react/lib/Breadcrumb';
 import formatMessage from 'format-message';
 import { globalHistory, RouteComponentProps } from '@reach/router';
@@ -104,6 +104,9 @@ const DesignPage: React.FC<RouteComponentProps<{ dialogId: string; projectId: st
     setectAndfocus,
     updateDialog,
     clearUndoHistory,
+    createDialogBegin,
+    exportToZip,
+    onboardingAddCoachMarkRef,
   } = actions;
   const { location, dialogId } = props;
   const params = new URLSearchParams(location?.search);
@@ -200,80 +203,164 @@ const DesignPage: React.FC<RouteComponentProps<{ dialogId: string; projectId: st
     }
   };
 
-  const nodeOperationAvailable = Array.isArray(visualEditorSelection) && visualEditorSelection.length > 0;
+  const { actionSelected, showDisableBtn, showEnableBtn } = useMemo(() => {
+    const actionSelected = Array.isArray(visualEditorSelection) && visualEditorSelection.length > 0;
+    if (!actionSelected) {
+      return {};
+    }
+    const selectedActions = visualEditorSelection.map((id) => get(currentDialog?.content, id));
+    const showDisableBtn = selectedActions.some((x) => get(x, 'disabled') !== true);
+    const showEnableBtn = selectedActions.some((x) => get(x, 'disabled') === true);
+    return { actionSelected, showDisableBtn, showEnableBtn };
+  }, [visualEditorSelection]);
 
   const toolbarItems: IToolBarItem[] = [
     {
-      type: 'action',
-      text: formatMessage('Undo'),
-      buttonProps: {
-        iconProps: {
-          iconName: 'Undo',
-        },
-        onClick: () => actions.undo(),
-      },
+      type: 'dropdown',
+      text: formatMessage('Add'),
       align: 'left',
-      disabled: !undoHistory.canUndo(),
+      dataTestid: 'AddFlyout',
+      buttonProps: {
+        iconProps: { iconName: 'Add' },
+      },
+      menuProps: {
+        items: [
+          {
+            'data-testid': 'FlyoutNewDialog',
+            key: 'adddialog',
+            text: formatMessage('Add new dialog'),
+            onClick: () => {
+              createDialogBegin([], onCreateDialogComplete);
+            },
+          },
+          {
+            'data-testid': 'FlyoutNewTrigger',
+            key: 'addtrigger',
+            text: formatMessage(`Add new trigger on {displayName}`, {
+              displayName: currentDialog?.displayName ?? '',
+            }),
+            onClick: () => {
+              openNewTriggerModal();
+            },
+          },
+        ],
+      },
     },
     {
-      type: 'action',
-      text: formatMessage('Redo'),
-      buttonProps: {
-        iconProps: {
-          iconName: 'Redo',
-        },
-        onClick: () => actions.redo(),
-      },
+      type: 'dropdown',
+      text: formatMessage('Edit'),
       align: 'left',
-      disabled: !undoHistory.canRedo(),
+      dataTestid: 'EditFlyout',
+      buttonProps: {
+        iconProps: { iconName: 'Edit' },
+      },
+      menuProps: {
+        items: [
+          {
+            key: 'edit.undo',
+            text: formatMessage('Undo'),
+            disabled: !undoHistory.canUndo(),
+            onClick: () => {
+              actions.undo();
+            },
+          },
+          {
+            key: 'edit.redo',
+            text: formatMessage('Redo'),
+            disabled: !undoHistory.canRedo(),
+            onClick: () => {
+              actions.redo();
+            },
+          },
+          {
+            key: 'edit.cut',
+            text: formatMessage('Cut'),
+            disabled: !actionSelected,
+            onClick: () => {
+              VisualEditorAPI.cutSelection();
+            },
+          },
+          {
+            key: 'edit.copy',
+            text: formatMessage('Copy'),
+            disabled: !actionSelected,
+            onClick: () => {
+              VisualEditorAPI.copySelection();
+            },
+          },
+          {
+            key: 'edit.move',
+            text: formatMessage('Move'),
+            disabled: !actionSelected,
+            onClick: () => {
+              VisualEditorAPI.moveSelection();
+            },
+          },
+          {
+            key: 'edit.delete',
+            text: formatMessage('Delete'),
+            disabled: !actionSelected,
+            onClick: () => {
+              VisualEditorAPI.deleteSelection();
+            },
+          },
+        ],
+      },
     },
     {
-      type: 'action',
-      text: formatMessage('Cut'),
-      buttonProps: {
-        iconProps: {
-          iconName: 'Cut',
-        },
-        onClick: () => VisualEditorAPI.cutSelection(),
-      },
+      type: 'dropdown',
+      text: formatMessage('Disable'),
       align: 'left',
-      disabled: !nodeOperationAvailable,
+      disabled: !actionSelected,
+      buttonProps: {
+        iconProps: { iconName: 'RemoveOccurrence' },
+      },
+      menuProps: {
+        items: [
+          {
+            key: 'disable',
+            text: formatMessage('Disable'),
+            disabled: !showDisableBtn,
+            onClick: () => {
+              VisualEditorAPI.disableSelection();
+            },
+          },
+          {
+            key: 'enable',
+            text: formatMessage('Enable'),
+            disabled: !showEnableBtn,
+            onClick: () => {
+              VisualEditorAPI.enableSelection();
+            },
+          },
+        ],
+      },
     },
     {
-      type: 'action',
-      text: formatMessage('Copy'),
-      buttonProps: {
-        iconProps: {
-          iconName: 'Copy',
-        },
-        onClick: () => VisualEditorAPI.copySelection(),
-      },
+      type: 'dropdown',
+      text: formatMessage('Export'),
       align: 'left',
-      disabled: !nodeOperationAvailable,
-    },
-    {
-      type: 'action',
-      text: formatMessage('Move'),
       buttonProps: {
-        iconProps: {
-          iconName: 'Share',
-        },
-        onClick: () => VisualEditorAPI.moveSelection(),
+        iconProps: { iconName: 'OpenInNewWindow' },
       },
-      align: 'left',
-      disabled: !nodeOperationAvailable,
-    },
-    {
-      type: 'action',
-      text: formatMessage('Delete'),
-      buttonProps: {
-        iconProps: {
-          iconName: 'Delete',
-        },
-        onClick: () => VisualEditorAPI.deleteSelection(),
+      menuProps: {
+        items: [
+          {
+            key: 'zipexport',
+            text: formatMessage('Export assets to .zip'),
+            onClick: () => {
+              exportToZip({ projectId });
+            },
+          },
+          {
+            key: 'exportAsSkill',
+            text: formatMessage('Export as skill'),
+            onClick: () => {
+              setExportSkillModalVisible(true);
+            },
+          },
+        ],
       },
-      align: 'left',
-      disabled: !nodeOperationAvailable,
     },
     {
       type: 'element',
@@ -396,6 +483,9 @@ const DesignPage: React.FC<RouteComponentProps<{ dialogId: string; projectId: st
       }
     }
   }
+  const addNewBtnRef = useCallback((addNew) => {
+    onboardingAddCoachMarkRef({ addNew });
+  }, []);
 
   if (!dialogId) {
     return <LoadingSpinner />;
@@ -413,13 +503,14 @@ const DesignPage: React.FC<RouteComponentProps<{ dialogId: string; projectId: st
           onSelect={handleSelect}
         />
         <div css={contentWrapper} role="main">
-          <ToolBar
-            currentDialog={currentDialog}
-            openNewTriggerModal={openNewTriggerModal}
-            showSkillManifestModal={() => setExportSkillModalVisible(true)}
-            toolbarItems={toolbarItems}
-            onCreateDialogComplete={onCreateDialogComplete}
-          />
+          <div css={{ position: 'relative' }} data-testid="DesignPage-ToolBar">
+            <span
+              ref={addNewBtnRef}
+              css={{ width: 120, height: '100%', position: 'absolute', left: 0, visibility: 'hidden' }}
+              data-testid="CoachmarkRef-AddNew"
+            />
+            <ToolBar toolbarItems={toolbarItems} />
+          </div>
           <Conversation css={editorContainer}>
             <div css={editorWrapper}>
               <div aria-label={formatMessage('Authoring canvas')} css={visualPanel} role="region">
